requirements.txt
pandas==2.3.2
Requests==2.32.5
python-dotenv==1.1.1

--extra-index-url http://packages.radioanalyzerserver.dk/simple
--trusted-host packages.radioanalyzerserver.dk
ra_autorun==2.0.6
ra_datetime_helper==2.2.6
ra_instance_spinner==2.0.11
ra_mysql_package==4.0.8
ra_radio_helper==2.1.9
ra_service_helper==4.0.43

